Given the following testing database configuration:
   - Installed modules: stock,stock_account,purchase,sale_management, account,account_accountant,l10n_generic_coa
   - Anglo-saxon accounting turned on
   - All product categories' property_valuation set to real_time

The following steps:
 - Create 80 storable products
 - Purchase each product to have enough in stock
 - Create and Confirm an SO with one SOL by product
 - Deliver products
 - Create Invoice (should have one line by SOL)
 - Post Invoice

lead to a significant slowdown. Up-to 5min30 to post the whole invoice.

A possible way to speed that up, at least in some cases, is to group the lines in `account_move_line._sync_invoice` by amount_currency and currency_fields. And then call `__set__` on all the lines with the same key at the same time, instead of one by one in the for loop.

In the best cases, this can lead to a speed up of factor 10-20.

#### Speedup
Timing of posting the invoice w.r.t. the number of unique amount currency, with constant currency fields.

| Number of Lines | Unique amount currency | Before PR | After PR |
|:----------------------:|:--------------------------------:|:--------------:|:-----------:|
| 80 | 80 | 5min30s | 5min36s |
| 80 | 40 | 5min36s | 3min |
| 80 | 20 | 5min36s | 1min30s |
| 80 | 5 | 5 min 30s | 32.76s |
| 80 | 1 | 5 min 30s| 16.82s |

Check for perf regression for small invoices with different amount currency:
- invoice with `|line_ids| = 1`: 271ms -> 293ms
- invoice with `|line_ids| = 3`: 469ms -> 372ms
- invoice with `|line_ids| = 5`: 526ms -> 608ms

The customer database linked to this ticket has on average 83 lines by invoice and 4 unique amount currency
so their expected speed up factor should be around 10.

opw-3135002



---
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr