@Aurelienvd 2023-02-07
@odoo/rd-accounting Hello, could someone have a look at this PR when they have the time ? Thank you !