@Aurelienvd 2023-02-07
@william-andre 

Actually the commit that you made and was merged yesterday https://github.com/odoo/odoo/commit/b3d60185 gave
a huge performance boost to this issue. I'm not really sure why but it's now faster than with this PR's fix so I guess I can close it 😅 
> 
> ```python
>     def test_invoice_perf(self):
>         import time
>         start = time.time()
>         move = self.env['account.move'].create({
>             'move_type': 'out_invoice',
>             'partner_id': self.partner_a.id,
>             'currency_id': self.company_data['currency'].id,
>             'invoice_date': '2020-02-02',
>             'invoice_line_ids': [
>                 Command.create({
>                     'price_unit': i,
>                     'name': f"{i}",
>                 })
>                 for i in range(400)
>             ]
>         })
>         move.action_post()
>         print(time.time() - start)
> ```
> 
> It prints `4.5630` (for 400 lines with different amounts)
> 
> Isn't this an issue in the client's code?

For an invoice with 200 lines, it still takes around 1min40s to post in a testing v16 database with standard modules. Is the method `_stock_account_anglo_saxon_reconcile_valuation` called in your example here ? 

Here is a flamegraph for this invoice posting in case you see something obvious that I don't: 

![16 0_invoice_post](https://user-images.githubusercontent.com/9673274/217265243-110aa9cc-02f6-4187-b62b-5e43b5a2b1cb.svg)