@william-andre 2023-02-08
I had a suspicion that that commit might have had an impact when seeing that you were setting the env `skip_invoice_line_sync` everywhere indeed, but I wasn't sure since it is not obvious that it prevents the recursion of that part of the code.