@Levizar 2023-02-08
Just passed by and saw the flame graph.
I saw that one of the longest function triggered was the creation of the full_reconcile_id.
Wouldn't that be faster if those creations were batched instead of being done each time in the for loop?